finished UI refactor of the codebase. renamed multiple files and structs.
