+++ STEP Navigation/Views/ARViewContainer.swift
@@ -7,16 +7,16 @@
 //  Wrap a ARView (UIKit) with a SwiftUI View struct
 //
 
-import Foundation
-import SwiftUI
-import ARKit
-
-struct ARViewContainer: UIViewRepresentable {
-    
-    func makeUIView(context: Context) -> ARSCNView {
-        return PositioningModel.shared.arView
-    }
-    
-    func updateUIView(_ uiView: ARSCNView, context: Context) {}
-    
-}
+//import Foundation
+//import SwiftUI
+//import ARKit
+//
+//struct ARViewContainer: UIViewRepresentable {
+//
+//    func makeUIView(context: Context) -> ARSCNView {
+//        return PositioningModel.shared.arView
+//    }
+//
+//    func updateUIView(_ uiView: ARSCNView, context: Context) {}
+//
+//}
