+++ STEP Navigation/Views/AnchorDetailView.swift
@@ -8,8 +8,10 @@
 import SwiftUI
 import CoreLocation
 
-struct AnchorDetailView: View {
+struct AnchorDetailView<Destination: View>: View {
     let anchorDetails: LocationDataModel
+    let buttonLabel: String
+    let buttonDestination: () -> Destination
     
     var body: some View {
         VStack {
@@ -20,15 +22,22 @@ struct AnchorDetailView: View {
                     .padding(.top)
             }
             Spacer()
-            if anchorDetails.getAnchorType().isIndoors {
-                SmallButtonComponent_NavigationLink(destination: {
-                    ChooseStartAnchorView(destinationAnchorDetails: anchorDetails)
-                                }, label: "Find Start Anchor")
-            } else {
-                SmallButtonComponent_NavigationLink(destination: {
-                    NavigatingView(startAnchorDetails: nil, destinationAnchorDetails: anchorDetails)
-                                }, label: "Navigate")
-            }
+            SmallButtonComponent_NavigationLink(destination: buttonDestination, label: buttonLabel)
+            
+//            if anchorDetails.getAnchorType().isIndoors {
+//                SmallButtonComponent_NavigationLink(destination: {
+//                    StartAnchorListView(destinationAnchorDetails: anchorDetails)
+//                                }, label: "Find Start Anchor")
+//            } else {
+//                SmallButtonComponent_NavigationLink(destination: {
+//                    NavigatingView(startAnchorDetails: nil, destinationAnchorDetails: anchorDetails)
+//                                }, label: "Navigate")
+//
+//                SmallButtonComponent_NavigationLink(destination: {
+//                    NavigatingView(startAnchorDetails: chosenStart, destinationAnchorDetails: destinationAnchorDetails!)
+//                                }, label: "Navigate")
+//
+//            }
         }
     }
 }
